fix(posix/vfs): 文件类 POSIX 上板暴露的两个真问题
上板（F427 开发槽 3）自检从 25ok/1FAIL 做到 32ok/0FAIL，途中暴露两处：

1) fstat 对按路径打开的字符设备回 -1
   /dev/<name> 是 VFS 句柄，fstat 对 VFS 一律 lseek(SEEK_END) 求大小，
   字符设备没有位置可 seek，于是 -1，与 sys/stat.h 的文档矛盾。
   改成 open 时先 svcrt_path_stat 缓存 S_IFMT 类型位进 fd 槽，fstat 按缓存
   回答，只有普通文件才 seek 求长度。

2) devfs 把内核设备句柄存进 int16_t，高位标志被截掉
   句柄是 (槽号 | SVCRT_DEV_HANDLE_FLAG 0x01200000)，截断后内核按句柄非法
   回 -1。修复在 ark_vfs 源仓库（6bb456a），此处是同步副本。
   另：一次写 0 字节不是错误——发送环满时字符设备允许短写，试验件改为按字节
   重试（有上限），33 字节全送到。

自检覆盖表、编译烧录命令与上板结论见 example/stm32f427/app_sdk/FS_DEMO/README.md
与 docs/POSIX与Windows兼容层说明.md 第 11 节。

+++ kernelsrc/components/ark_vfs/port/ark_vfs_devfs.c
@@ -42,7 +42,7 @@ typedef int ark_vfs_devfs_disabled_tu;
 struct ark_vfs_file {
     uint8_t  used;
     uint8_t  is_dir;
-    int16_t  dev;                   /* svcrt_dev handle, or -1          */
+    int32_t  dev;                   /* svcrt_dev handle, or -1          */
     uint32_t scan;                  /* readdir cursor, slot index       */
 };
 
@@ -159,7 +159,7 @@ static int32_t devfs_open(struct ark_vfs_mnt *mnt, const char *rel, int flags,
         return ARK_E_FULL;
     }
     h->is_dir = 0u;
-    h->dev    = (int16_t)dev;
+    h->dev    = dev;
     *out = h;
     return ARK_VFS_OK;
 }
